+++ src/OpenSimCreator/Tabs/MeshImporterTab.cpp
@@ -13,6 +13,7 @@
 #include <oscar/Bindings/GlmHelpers.hpp>
 #include <oscar/Bindings/ImGuiHelpers.hpp>
 #include <oscar/Bindings/ImGuizmoHelpers.hpp>
+#include <oscar/Formats/CSV.hpp>
 #include <oscar/Formats/OBJ.hpp>
 #include <oscar/Formats/STL.hpp>
 #include <oscar/Graphics/Color.hpp>
@@ -53,7 +54,11 @@
 #include <oscar/Utils/Spsc.hpp>
 #include <oscar/Utils/StringHelpers.hpp>
 #include <oscar/Utils/UID.hpp>
+#include <oscar/Utils/VariantHelpers.hpp>
 #include <oscar/Widgets/LogViewer.hpp>
+#include <oscar/Widgets/Popup.hpp>
+#include <oscar/Widgets/PopupManager.hpp>
+#include <oscar/Widgets/StandardPopup.hpp>
 #include <OscarConfiguration.hpp>
 
 #include <glm/mat3x3.hpp>
@@ -6132,6 +6137,339 @@ namespace
     };
 }
 
+// popups
+namespace
+{
+    class ImportStationsFromCSVPopup final : public osc::StandardPopup {
+    public:
+        ImportStationsFromCSVPopup(
+            std::string_view popupName_,
+            std::shared_ptr<SharedData> shared_) :
+
+            StandardPopup{popupName_},
+            m_Shared{std::move(shared_)}
+        {
+            setModal(true);
+        }
+
+    private:
+        struct StationDefinedInGround final {
+            std::string name;
+            glm::vec3 location;
+        };
+
+        struct StationsDefinedInGround final {
+            std::vector<StationDefinedInGround> rows;
+        };
+
+        struct ImportedCSVData final {
+            std::filesystem::path sourceDataPath;
+            std::variant<StationsDefinedInGround> parsedData;
+        };
+
+        struct CSVImportError final {
+            std::filesystem::path userSelectedPath;
+            std::string message;
+        };
+
+        using CSVImportResult = std::variant<ImportedCSVData, CSVImportError>;
+
+        struct RowParseError final {
+            size_t lineNum;
+            std::string errorMsg;
+        };
+
+        static std::variant<StationDefinedInGround, RowParseError> TryParseColumns(
+            size_t lineNum,
+            nonstd::span<std::string const> columnsText)
+        {
+            if (columnsText.size() < 4)
+            {
+                return RowParseError{lineNum, "too few columns in this row (expecting at least 4)"};
+            }
+
+            std::string const stationName = columnsText[0];
+
+            std::optional<float> const maybeX = osc::FromCharsStripWhitespace(columnsText[1]);
+            if (!maybeX)
+            {
+                return RowParseError{lineNum, "cannot parse X as a number"};
+            }
+
+            std::optional<float> const maybeY = osc::FromCharsStripWhitespace(columnsText[2]);
+            if (!maybeY)
+            {
+                return RowParseError{lineNum, "cannot parse Y as a number"};
+            }
+
+            std::optional<float> const maybeZ = osc::FromCharsStripWhitespace(columnsText[3]);
+            if (!maybeZ)
+            {
+                return RowParseError{lineNum, "cannot parse Z as a number"};
+            }
+
+            glm::vec3 const locationInGround = {*maybeX, *maybeY, *maybeZ};
+
+            return StationDefinedInGround{stationName, locationInGround};
+        }
+
+        static std::string to_string(RowParseError const& e)
+        {
+            std::stringstream ss;
+            ss << "line " << e.lineNum << ": " << e.errorMsg;
+            return std::move(ss).str();
+        }
+
+        static bool IsWhitespaceRow(nonstd::span<std::string const> cols)
+        {
+            return cols.size() == 1;
+        }
+
+        static CSVImportResult TryReadCSVInput(std::filesystem::path const& path, std::istream& input)
+        {
+            // input must contain at least one (header) row
+            if (!osc::ReadCSVRow(input))
+            {
+                return CSVImportError{path, "cannot read a header row from the input (is the file empty?)"};
+            }
+
+            // then try to read each row as a data row, propagating errors
+            // accordingly
+
+            StationsDefinedInGround successfullyParsedStations;
+            std::optional<RowParseError> maybeParseError;
+            {
+                size_t lineNum = 1;
+                for (std::vector<std::string> row;
+                    !maybeParseError && osc::ReadCSVRowIntoVector(input, row);
+                    ++lineNum)
+                {
+                    if (IsWhitespaceRow(row))
+                    {
+                        continue;  // skip
+                    }
+
+                    // else: try parsing the row as a data row
+                    std::visit(osc::Overload
+                    {
+                        [&successfullyParsedStations](StationDefinedInGround const& success)
+                        {
+                            successfullyParsedStations.rows.push_back(success);
+                        },
+                        [&maybeParseError](RowParseError const& fail)
+                        {
+                            maybeParseError = fail;
+                        },
+                    }, TryParseColumns(lineNum, row));
+                }
+            }
+
+            if (maybeParseError)
+            {
+                return CSVImportError{path, to_string(*maybeParseError)};
+            }
+            else
+            {
+                return ImportedCSVData{path, std::move(successfullyParsedStations)};
+            }
+        }
+
+        static CSVImportResult TryReadCSVFile(std::filesystem::path const& path)
+        {
+            std::ifstream f{path};
+            if (!f)
+            {
+                return CSVImportError{path, "cannot open the provided file for reading"};
+            }
+            f.setf(std::ios_base::skipws);
+
+            return TryReadCSVInput(path, f);
+        }
+
+        void implDrawContent() final
+        {
+            drawHelpText();
+
+            ImGui::Dummy({0.0f, 0.25f*ImGui::GetTextLineHeight()});
+            ImGui::Separator();
+            ImGui::Dummy({0.0f, 0.5f*ImGui::GetTextLineHeight()});
+
+            if (m_MaybeImportResult.has_value())
+            {
+                std::visit(osc::Overload
+                {
+                    [this](ImportedCSVData const& data) { drawLoadedFileState(data); },
+                    [this](CSVImportError const& error) { drawErrorLoadingFileState(error); }
+                }, *m_MaybeImportResult);
+            }
+            else
+            {
+                drawSelectInitialFileState();
+            }
+        }
+
+        void drawHelpText()
+        {
+            ImGui::TextWrapped("Use this tool to import CSV data containing 3D locations as stations into the mesh importer scene. The CSV file should contain");
+            ImGui::Bullet();
+            ImGui::TextWrapped("A header row of four columns, ideally labelled 'name', 'x', 'y', and 'z'");
+            ImGui::Bullet();
+            ImGui::TextWrapped("Data rows containing four columns: name (string), x (number), y (number), and z (number)");
+            ImGui::TextWrapped("Example Input:");
+            ImGui::Indent();
+            ImGui::TextWrapped("name,x,y,z\nstationatground,0,0,0\nstation2,1.53,0.2,1.7\nstation3,3.0,2.0,0.0\netc.");
+            ImGui::Unindent();
+        }
+
+        void drawSelectInitialFileState()
+        {
+            if (ImGui::Button("Select File"))
+            {
+                actionTryPromptingUserForCSVFile();
+            }
+
+            ImGui::Dummy({0.0f, 0.25f*ImGui::GetTextLineHeight()});
+            ImGui::Separator();
+            ImGui::Dummy({0.0f, 0.5f*ImGui::GetTextLineHeight()});
+
+            drawDisabledOkCancelButtons("Cannot continue: nothing has been imported (select a file first)");
+        }
+
+        void drawErrorLoadingFileState(CSVImportError const& error)
+        {
+            ImGui::Text("Error loading %s: %s ", error.userSelectedPath.string().c_str(), error.message.c_str());
+            if (ImGui::Button("Try Again (Select File)"))
+            {
+                actionTryPromptingUserForCSVFile();
+            }
+
+            ImGui::Dummy({0.0f, 0.25f*ImGui::GetTextLineHeight()});
+            ImGui::Separator();
+            ImGui::Dummy({0.0f, 0.5f*ImGui::GetTextLineHeight()});
+
+            drawDisabledOkCancelButtons("Cannot continue: there is an error in the imported data (try again)");
+        }
+
+        void drawDisabledOkCancelButtons(osc::CStringView disabledReason)
+        {
+            ImGui::BeginDisabled();
+            ImGui::Button("OK");
+            ImGui::EndDisabled();
+            if (ImGui::IsItemHovered(ImGuiHoveredFlags_AllowWhenDisabled))
+            {
+                osc::DrawTooltipBodyOnly(disabledReason);
+            }
+            ImGui::SameLine();
+            if (ImGui::Button("Cancel"))
+            {
+                close();
+            }
+        }
+
+        void drawLoadedFileState(ImportedCSVData const& result)
+        {
+            std::visit(osc::Overload
+            {
+                [this, &result](StationsDefinedInGround const& data) { drawLoadedFileStateData(result, data); },
+            }, result.parsedData);
+
+            ImGui::Dummy({0.0f, 0.25f*ImGui::GetTextLineHeight()});
+            ImGui::Separator();
+            ImGui::Dummy({0.0f, 0.5f*ImGui::GetTextLineHeight()});
+
+            if (ImGui::Button("OK"))
+            {
+                actionAttachResultToModelGraph(result);
+                close();
+            }
+            ImGui::SameLine();
+            if (ImGui::Button("Cancel"))
+            {
+                close();
+            }
+        }
+
+        void drawLoadedFileStateData(ImportedCSVData const& result, StationsDefinedInGround const& data)
+        {
+            ImGui::Text("%s (%zu data rows)", result.sourceDataPath.string().c_str(), data.rows.size());
+
+            if (ImGui::BeginTable("##importtable", 4, ImGuiTableFlags_ScrollY, {0.0f, 10.0f*ImGui::GetTextLineHeight()}))
+            {
+                ImGui::TableSetupColumn("Name");
+                ImGui::TableSetupColumn("X");
+                ImGui::TableSetupColumn("Y");
+                ImGui::TableSetupColumn("Z");
+                ImGui::TableHeadersRow();
+
+                int id = 0;
+                for (StationDefinedInGround const& row : data.rows)
+                {
+                    ImGui::PushID(id++);
+                    ImGui::TableNextRow();
+                    int column = 0;
+                    ImGui::TableSetColumnIndex(column++);
+                    ImGui::TextUnformatted(row.name.c_str());
+                    ImGui::TableSetColumnIndex(column++);
+                    ImGui::Text("%f", row.location.x);
+                    ImGui::TableSetColumnIndex(column++);
+                    ImGui::Text("%f", row.location.y);
+                    ImGui::TableSetColumnIndex(column++);
+                    ImGui::Text("%f", row.location.z);
+                    ImGui::PopID();
+                }
+
+                ImGui::EndTable();
+            }
+
+            if (ImGui::Button("Select Different File"))
+            {
+                actionTryPromptingUserForCSVFile();
+            }
+        }
+
+        void actionTryPromptingUserForCSVFile()
+        {
+            if (auto path = osc::PromptUserForFile("csv"))
+            {
+                m_MaybeImportResult = TryReadCSVFile(*path);
+            }
+        }
+
+        void actionAttachResultToModelGraph(ImportedCSVData const& result)
+        {
+            std::visit(osc::Overload
+            {
+                [this, &result](StationsDefinedInGround const& data) { actionAttachStationsInGroundToModelGraph(result, data); },
+            }, result.parsedData);
+        }
+
+        void actionAttachStationsInGroundToModelGraph(
+            ImportedCSVData const& result,
+            StationsDefinedInGround const& data)
+        {
+            CommittableModelGraph& undoable = m_Shared->UpdCommittableModelGraph();
+
+            ModelGraph& graph = undoable.UpdScratch();
+            for (StationDefinedInGround const& station : data.rows)
+            {
+                graph.AddEl<StationEl>(
+                    UIDT<StationEl>{},
+                    c_GroundID,
+                    station.location,
+                    station.name
+                );
+            }
+
+            std::stringstream ss;
+            ss << "imported " << result.sourceDataPath;
+            undoable.Commit(std::move(ss).str());
+        }
+
+        std::shared_ptr<SharedData> m_Shared;
+        std::optional<CSVImportResult> m_MaybeImportResult;
+    };
+}
+
 // mesh importer tab implementation
 class osc::MeshImporterTab::Impl final : public LayerHost {
 public:
@@ -6182,6 +6520,7 @@ public:
     void onMount()
     {
         App::upd().makeMainEventLoopWaiting();
+        m_PopupManager.onMount();
     }
 
     void onUnmount()
@@ -6321,6 +6660,9 @@ public:
 
         // draw contextual 3D modal (if there is one), else: draw standard 3D viewer
         DrawMainViewerPanelOrModal();
+
+        // draw any active popups over the scene
+        m_PopupManager.onDraw();
     }
 
 private:
@@ -8353,6 +8695,8 @@ private:
                 m_Shared->RequestNewMeshImporterTab();
             }
 
+            ImGui::Separator();
+
             if (ImGui::MenuItem(ICON_FA_FOLDER_OPEN " Import", "Ctrl+O"))
             {
                 m_Shared->OpenOsimFileAsModelGraph();
@@ -8370,6 +8714,20 @@ private:
                 m_Shared->ExportAsModelGraphAsOsimFile();
             }
             osc::DrawTooltipIfItemHovered("Export mesh impoter scene to osim", "Try to export the current mesh importer scene to an osim.\n\nBEWARE: the mesh importer scene may not map 1:1 onto an OpenSim model, so re-importing the scene *may* change a few things slightly. The main utility of this button is to try and save some progress in the mesh importer.");
+
+            ImGui::Separator();
+
+            if (ImGui::MenuItem(ICON_FA_FOLDER_OPEN " Import Stations from CSV"))
+            {
+                auto popup = std::make_shared<ImportStationsFromCSVPopup>(
+                    "Import Stations from CSV",
+                    m_Shared
+                );
+                popup->open();
+                m_PopupManager.push_back(std::move(popup));
+            }
+
+            ImGui::Separator();
 
             if (ImGui::MenuItem(ICON_FA_TIMES " Close", "Ctrl+W"))
             {
@@ -8500,6 +8858,9 @@ private:
         ImGuizmo::OPERATION op = ImGuizmo::TRANSLATE;
         ImGuizmo::MODE mode = ImGuizmo::WORLD;
     } m_ImGuizmoState;
+
+    // manager for active modal popups (importer popups, etc.)
+    PopupManager m_PopupManager;
 };